pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aml-open-framework"
version = "0.1.0"
description = "Spec-driven, audit-ready AML automation for regulated financial institutions."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [{ name = "AML Open Framework contributors" }]
keywords = ["aml", "compliance", "fintech", "regtech", "spec-driven"]
dependencies = [
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "jsonschema>=4.21",
    "duckdb>=0.10",
    "jinja2>=3.1",
    "typer>=0.12",
    "rich>=13.7",
    "faker>=24.0",
    "python-dateutil>=2.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[project.scripts]
aml = "aml_framework.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
aml_framework = ["templates/*.sql.j2", "templates/*.md.j2"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"

[tool.ruff]
line-length = 100
target-version = "py310"
